+++ webclient/src/main.js
@@ -1,5 +1,5 @@
 import Vue from 'vue';
-import store from './assets/js/vuex/index';
+import store from './js/vuex/index';
 
 import App from './App.vue';
 Vue.use(store);
